1. Implement EV+MxD.
2. Implement Pre/Post Plus operations. The first operand is an EV+MxD and the second operand is an EV+MDD. The result is an EV+MxD of the From/To states in the first operand plus the second operand.


git-svn-id: svn://svn.code.sf.net/p/meddly/code/branches/cjiang-ev@785 14966d26-f10a-4722-9547-d5dcc201c352

+++ src/meddly_expert.h
@@ -515,7 +515,8 @@ class MEDDLY::unpacked_node {
     static unpacked_node* newRedundant(const expert_forest *f, int k, float ev, node_handle node, bool full);
 
     static unpacked_node* newIdentity(const expert_forest *f, int k, int i, node_handle node, bool full);
-    static unpacked_node* newIdentity(const expert_forest *f, int k, int i, int ev, node_handle node, bool full);
+    static unpacked_node* newIdentity(const expert_forest *f, int k, int i, long ev, node_handle node, bool full);
+//    static unpacked_node* newIdentity(const expert_forest *f, int k, int i, int ev, node_handle node, bool full);
     static unpacked_node* newIdentity(const expert_forest *f, int k, int i, float ev, node_handle node, bool full);
 
     static unpacked_node* newFull(const expert_forest *f, int level, int tsz);
